corregir problema con el C++ y la raiz cuadrada de un entero
Subversion-commit: r996

+++ src/eps/impls/lanczos/lanczos.c
@@ -266,7 +266,7 @@ static PetscErrorCode EPSPeriodicLanczos(EPS eps,PetscScalar *T,Vec *V,int k,int
     } else if (j>1) {
       ierr = VecMDot(j-1,f,eps->V,omega);CHKERRQ(ierr);
       for (i=0;i<j-1 && !reorthog;i++) 
-	if (PetscAbsScalar(omega[i]) > PETSC_SQRT_MACHINE_EPSILON/sqrt(j)) {
+	if (PetscAbsScalar(omega[i]) > PETSC_SQRT_MACHINE_EPSILON/sqrt((double)j)) {
 	  reorthog = PETSC_TRUE;
 	}
       if (reorthog) {
